+++ tests/BundlesTest.php
@@ -2,58 +2,14 @@
 
 namespace Tatter\Frontend\Bundles;
 
-use CodeIgniter\Publisher\Publisher;
-use Tatter\Frontend\FrontendBundle;
-use Tests\Support\TestCase;
+use Tatter\Frontend\Test\BundlesTestCase;
 
 /**
  * @internal
  */
-final class BundlesTest extends TestCase
+final class BundlesTest extends BundlesTestCase
 {
-    private $didPublish = false;
-
-    /**
-     * Publishes all files once so they are
-     * available for bundles.
-     */
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        // Make sure everything is published
-        if (! $this->didPublish) {
-            foreach (Publisher::discover() as $publisher) {
-                $publisher->publish();
-            }
-
-            $this->didPublish = true;
-        }
-    }
-
-    /**
-     * @dataProvider bundleProvider
-     *
-     * @param class-string<FrontendBundle> $class
-     * @param string[]                     $expectedHeadFiles
-     * @param string[]                     $expectedBodyFiles
-     */
-    public function testBundlesFiles(string $class, array $expectedHeadFiles, array $expectedBodyFiles): void
-    {
-        $bundle = new $class();
-        $head   = $bundle->head();
-        $body   = $bundle->body();
-
-        foreach ($expectedHeadFiles as $file) {
-            $this->assertStringContainsString($file, $head);
-        }
-
-        foreach ($expectedBodyFiles as $file) {
-            $this->assertStringContainsString($file, $body);
-        }
-    }
-
-    public function bundleProvider()
+    public function bundleProvider(): array
     {
         return [
             [
